Moved Publication models to own sub directory
