+++ aspnetcore/src/Api/Models/Publication/Author.cs
@@ -1,4 +1,4 @@
-﻿namespace Api.Models
+﻿namespace Api.Models.Publication
 {
     public class Author
     {
