+++ aspnetcore/src/Api/Models/Publication/OrganizationUnit.cs
@@ -1,6 +1,6 @@
 ﻿using Nest;
 
-namespace Api.Models
+namespace Api.Models.Publication
 {
     public class OrganizationUnit
     {
