+++ aspnetcore/src/Api/Models/Publication/PeerReviewed.cs
@@ -1,4 +1,4 @@
-﻿namespace Api.Models
+﻿namespace Api.Models.Publication
 {
     public class PeerReviewed
     {
